requirements.txt
flask
requests
replicate
Pillow
